+++ sdk/security/keyvault/azkeys/CHANGELOG.md
@@ -6,7 +6,11 @@
 - New enum type `JSONWebKeyWrapAlgorithm` with values `JSONWebKeyWrapAlgorithmA128KW`, `JSONWebKeyWrapAlgorithmA128KWPAD`, `JSONWebKeyWrapAlgorithmA192KW`, `JSONWebKeyWrapAlgorithmA192KWPAD`, `JSONWebKeyWrapAlgorithmA256KW`, `JSONWebKeyWrapAlgorithmA256KWPAD`, `JSONWebKeyWrapAlgorithmCKMAESKEYWRAP`, `JSONWebKeyWrapAlgorithmCKMAESKEYWRAPPAD`, `JSONWebKeyWrapAlgorithmRSAOAEP256`
 - New function `*Client.SecureUnwrapKey(ctx context.Context, name string, version string, parameters SecureKeyUnWrapOperationParameters, options *SecureUnwrapKeyOptions) (SecureUnwrapKeyResponse, error)`
 - New function `*Client.SecureWrapKey(ctx context.Context, name string, version string, parameters SecureKeyWrapOperationParameters, options *SecureWrapKeyOptions) (SecureWrapKeyResponse, error)`
+- New function `NewSecureUnwrapKeyParameters(algorithm JSONWebKeyWrapAlgorithm, value []byte, targetAttestationToken string) SecureUnwrapKeyParameters`
+- New function `NewSecureWrapKeyParameters(algorithm JSONWebKeyWrapAlgorithm) SecureWrapKeyParameters`
 - New struct `ExternalKey`
+- New type alias `SecureUnwrapKeyParameters` for `SecureKeyUnWrapOperationParameters`
+- New type alias `SecureWrapKeyParameters` for `SecureKeyWrapOperationParameters`
 - New struct `SecureKeyOperationResult`
 - New struct `SecureKeyUnWrapOperationParameters`
 - New struct `SecureKeyWrapOperationParameters`